Fixes bugs and adds processing of unpaired reads

+++ batch_processing/2_filter_pe.sh
@@ -10,17 +10,26 @@ BOWTIE_DB=/labcommon/db/bowtie2/grch38
 SAMPLE_ID=$1
 IN_DIR=$2
 
+OUT_FILE=${SAMPLE_ID}_pe.fq.gz
+
+if [ -e ${OUT_FILE} ]; then
+	echo "File ${OUT_FILE} already exists. Delete file to regenerate."
+	exit 0;
+fi
+
 #Create named PIPEs
 mkfifo R1_PIPE R2_PIPE
 
+mkdir -p log
+
 #Requires cutadapt
 cutadapt --cut=3 -g ^GCCGGAGCTCTGCAGATATC -g ^GGAGCTCTGCAGATATC --no-indels --error-rate=0.1 \
-	-o R1_PIPE $(IN_DIR)/*_val_1.fq.gz 2>&1 > overhangs_R1.log &
+	-o R1_PIPE ${IN_DIR}/*_val_1.fq.gz 2>&1 > log/overhangs_R1.log &
 
 cutadapt --cut=3 -g ^GCCGGAGCTCTGCAGATATC -g ^GGAGCTCTGCAGATATC --no-indels --error-rate=0.1 \
-	-o R2_PIPE $(IN_DIR)/*_val_2.fq.gz 2>&1 > overhangs_R2.log &
+	-o R2_PIPE ${IN_DIR}/*_val_2.fq.gz 2>&1 > log/overhangs_R2.log &
 
-bowtie2 --local --very-sensitive-local -t -p 8 -x ${BOWTIE_DB} -1 R1_PIPE -2 R2_PIPE | tee >(samtools view -hSb - | samtools flgstat - > bowtie_pe.flagstat ) | samtools view -hSb -f12 -F256 - | samtools fastq - | gzip > ${SAMPLE_ID}_pe.fq.gz
+bowtie2 --local --very-sensitive-local -t -p 8 -x ${BOWTIE_DB} -1 R1_PIPE -2 R2_PIPE | tee >(samtools view -hSb - | samtools flagstat - > log/bowtie_pe.flagstat ) | samtools view -hSb -f12 -F256 - | samtools fastq - | gzip > ${OUT_FILE}
 
 #Remove named PIPEs
 rm R1_PIPE R2_PIPE
